requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
certifi==2022.6.15; python_full_version >= '3.6.0'
charset-normalizer==2.1.0; python_full_version >= '3.6.0'
commonmark==0.9.1
coverage==6.4.2
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django-test-plus==2.2.0
django==4.0.6
djangorestframework==3.13.1
faker==13.15.1
httpie==3.2.1
idna==3.3; python_version >= '3.5'
isort==5.10.1
multidict==6.0.2; python_version >= '3.7'
pygments==2.12.0; python_full_version >= '3.6.0'
pysocks==1.7.1
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.20.0
pytz==2022.1
requests-toolbelt==0.9.1
requests[socks]==2.28.1; python_version >= '3.7' and python_full_version < '4.0.0'
rich==12.5.1; python_full_version >= '3.6.3' and python_full_version < '4.0.0'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
urllib3==1.26.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_full_version < '4.0.0'
